+++ CHANGELOG.md
@@ -1,4 +1,4 @@
 For the documentation see the README on the main page.
 
-- added missing `amount:` in documentation
-- fixed exception in `hasItems` and `hasItemInHand` when no IA Item was found
+- validate amount with new variableNumber methods
+- fix using removed method in VariableNumber